Extract tests for ``pylint.pyreverse.utils`` into own test module.
